+++ src/sdc/help.py
@@ -1,7 +1,7 @@
 import os
 
 from seppl import OutputProducer, InputConsumer, classes_to_str, get_aliases, has_aliases
-from seppl.placeholders import PlaceholderSupporter, placeholder_help
+from seppl.variables import VariableSupporter, variable_help
 from sdc.registry import available_plugins
 from sdc.api import DefaultExtensionWriter
 
@@ -48,8 +48,8 @@ def generate_plugin_usage(plugin_name: str, help_format: str = HELP_FORMAT_TEXT,
         result = result.strip()
         result += "\n\n"
         result += plugin.format_help() + "\n"
-        if isinstance(plugin, PlaceholderSupporter):
-            result += "\n" + placeholder_help(markdown=False, obj=plugin) + "\n"
+        if isinstance(plugin, VariableSupporter):
+            result += "\n" + variable_help(markdown=False, obj=plugin) + "\n"
     elif help_format == HELP_FORMAT_MARKDOWN:
         suffix = ".md"
         result += "#"*heading_level + " " + plugin_name + "\n"
@@ -69,8 +69,8 @@ def generate_plugin_usage(plugin_name: str, help_format: str = HELP_FORMAT_TEXT,
         result += "```\n"
         result += plugin.format_help()
         result += "```\n"
-        if isinstance(plugin, PlaceholderSupporter):
-            result += "\n" + placeholder_help(markdown=True, obj=plugin) + "\n"
+        if isinstance(plugin, VariableSupporter):
+            result += "\n" + variable_help(markdown=True, obj=plugin) + "\n"
     else:
         raise Exception("Unhandled help format: %s" % help_format)