+++ src/sdc/reader/_nir.py
@@ -2,14 +2,14 @@ import argparse
 from typing import List, Iterable, Union
 
 from seppl.io import locate_files, DirectReader
-from seppl.placeholders import PlaceholderSupporter, placeholder_list
+from seppl.variables import VariableSupporter, variable_list
 from wai.logging import LOGGING_WARNING
 from wai.spectralio.nir import Reader as SReader
 
 from sdc.api import SpectralIOReader, Spectrum2D
 
 
-class NIRReader(SpectralIOReader, DirectReader, PlaceholderSupporter):
+class NIRReader(SpectralIOReader, DirectReader, VariableSupporter):
 
     def __init__(self, source: Union[str, List[str]] = None, source_list: Union[str, List[str]] = None,
                  resume_from: str = None, instrument: str = None, format: str = None, keep_format: bool = None,
@@ -84,8 +84,8 @@ class NIRReader(SpectralIOReader, DirectReader, PlaceholderSupporter):
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-i", "--input", type=str, help="Path to the NIR file(s) to read; glob syntax is supported; " + placeholder_list(obj=self), required=False, nargs="*")
-        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the NIR files to use; " + placeholder_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-i", "--input", type=str, help="Path to the NIR file(s) to read; glob syntax is supported; " + variable_list(obj=self), required=False, nargs="*")
+        parser.add_argument("-I", "--input_list", type=str, help="Path to the text file(s) listing the NIR files to use; " + variable_list(obj=self), required=False, nargs="*")
         parser.add_argument("--resume_from", type=str, help="Glob expression matching the file to resume from, e.g., '*/012345.nir'", required=False)
         parser.add_argument("--type_field", type=str, help="Code|Field1|Field2|Field3|ID|[sample_type]", required=False, default="Code")
         parser.add_argument("--id_field", type=str, help="ID|Field1|Field2|Field3|[prefix]", required=False, default="ID")