+++ src/sdc/writer/_asciixy.py
@@ -2,7 +2,7 @@ import argparse
 import os
 from typing import List
 
-from seppl.placeholders import placeholder_list, InputBasedPlaceholderSupporter
+from seppl.variables import InputBasedVariableSupporter, variable_list
 from seppl.io import DirectStreamWriter
 from wai.logging import LOGGING_WARNING
 from wai.spectralio.asciixy import Writer as SWriter
@@ -11,7 +11,7 @@ from kasperl.api import SplittableStreamWriter, make_list
 from sdc.api import Spectrum2D, SpectralIOWriter, DefaultExtensionWriter
 
 
-class ASCIIXYWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
+class ASCIIXYWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter, DefaultExtensionWriter, InputBasedVariableSupporter):
 
     def __init__(self, output_dir: str = None, separator: str = None,
                  split_names: List[str] = None, split_ratios: List[int] = None, split_group: str = None,
@@ -75,7 +75,7 @@ class ASCIIXYWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter
         :rtype: argparse.ArgumentParser
         """
         parser = super()._create_argparser()
-        parser.add_argument("-o", "--output", type=str, help="The directory to store the ASCII XY .txt files in. Any defined splits get added beneath there. " + placeholder_list(obj=self), required=False)
+        parser.add_argument("-o", "--output", type=str, help="The directory to store the ASCII XY .txt files in. Any defined splits get added beneath there. " + variable_list(obj=self), required=False)
         parser.add_argument("-s", "--separator", type=str, help="The separator to use for identifying X and Y columns.", required=False, default=";")
         return parser
 
@@ -139,7 +139,7 @@ class ASCIIXYWriter(SplittableStreamWriter, SpectralIOWriter, DirectStreamWriter
             raise Exception("No output directory specified!")
 
         for item in make_list(data):
-            sub_dir = self.session.expand_placeholders(self.output_dir)
+            sub_dir = self.session.expand_variables(self.output_dir)
             if self.splitter is not None:
                 split = self.splitter.next(item=item.spectrum_name)
                 sub_dir = os.path.join(sub_dir, split)