+++ src/sdc/writer/_nir.py
@@ -1,7 +1,7 @@
 import argparse
 from typing import List
 
-from seppl.placeholders import InputBasedPlaceholderSupporter
+from seppl.variables import InputBasedVariableSupporter
 from seppl.io import DirectBatchWriter
 from wai.logging import LOGGING_WARNING
 from wai.spectralio.nir import Writer as SWriter
@@ -10,7 +10,7 @@ from kasperl.api import SplittableBatchWriter, make_list
 from sdc.api import Spectrum2D, SpectralIOWriter, DefaultExtensionWriter
 
 
-class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, DefaultExtensionWriter, InputBasedPlaceholderSupporter):
+class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, DefaultExtensionWriter, InputBasedVariableSupporter):
 
     def __init__(self, output_file: str = None, instrument_name: str = None,
                  product_code: str = None, product_code_from_field: bool = None, client: str = None,
@@ -278,7 +278,7 @@ class NIRWriter(SplittableBatchWriter, SpectralIOWriter, DirectBatchWriter, Defa
         if self.output_file is None:
             raise Exception("No output file specified!")
 
-        output_file = self.session.expand_placeholders(self.output_file)
+        output_file = self.session.expand_variables(self.output_file)
         self.logger().info("Writing spectra to: %s" % output_file)
         self._writer.write([x.spectrum for x in data], output_file)